# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/edu-api/files/{name}": {
      "post": {
        "operationId": "FileSystemController_uploadFileToApp",
        "parameters": [],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["files"]
      }
    },
...
    "/edu-api/files/file/{appName}/{filename}": {
      "get": {
        "operationId": "FileSystemController_serveFile",
        "parameters": [{ "name": "appName", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["files"]
      }
    },
...
    "/edu-api/files/public/file/{appName}/{filename}": {
      "get": {
        "operationId": "FileSystemController_servePublicFile",
        "parameters": [{ "name": "appName", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["files"]
      }
    },
...
    "/edu-api/files": {
      "post": {
        "operationId": "FileSystemController_upload",
        "parameters": [],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["files"]
      }
    },
...
    "/edu-api/filesharing": {
      "get": {
        "operationId": "FilesharingController_getFilesAtPath",
        "parameters": [
          { "name": "type", "required": true, "in": "query", "schema": { "type": "string" } },
          { "name": "path", "required": true, "in": "query", "schema": { "type": "string" } },
          { "name": "share", "required": true, "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["filesharing"]
      },
      "post": {
        "operationId": "FilesharingController_createFileOrFolder",
        "parameters": [
          { "name": "path", "required": true, "in": "query", "schema": { "type": "string" } },
          { "name": "share", "required": true, "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["filesharing"]
      },
      "delete": {
        "operationId": "FilesharingController_deleteFile",
        "parameters": [{ "name": "share", "required": true, "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["filesharing"]
      },
      "patch": {
        "operationId": "FilesharingController_moveOrRenameResource",
        "parameters": [{ "name": "share", "required": true, "in": "query", "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "array", "items": { "type": "string" } } } }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["filesharing"]
      }
    },
...
    "/edu-api/wopi/files/{fileId}": {
      "get": {
        "operationId": "WopiController_checkFileInfo",
        "parameters": [
          { "name": "fileId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "access_token", "required": true, "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Wopi"]
      }
    },
    "/edu-api/wopi/files/{fileId}/contents": {
      "get": {
        "operationId": "WopiController_getFile",
        "parameters": [{ "name": "access_token", "required": true, "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Wopi"]
      },
      "post": {
        "operationId": "WopiController_putFile",
        "parameters": [{ "name": "access_token", "required": true, "in": "query", "schema": { "type": "string" } }],
        "responses": { "201": { "description": "" } },
        "tags": ["Wopi"]
      }
    },
...
    "/edu-api/surveys/files": {
      "post": {
        "operationId": "SurveysController_fileUpload",
        "parameters": [],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["surveys"]
      }
    },
...
    "/edu-api/surveys/answers/files/{userName}/{surveyId}/{questionId}/{filename}": {
      "get": {
        "operationId": "SurveysController_serveFileFromAnswer",
        "parameters": [
          { "name": "userName", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "surveyId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "questionId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "filename", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["surveys"]
      }
    },
...
    "/edu-api/surveys/files/{surveyId}/{questionId}/{filename}": {
      "get": {
        "operationId": "SurveysController_serveFile",
        "parameters": [
          { "name": "surveyId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "questionId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "filename", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["surveys"]
      }
    },
    "/edu-api/surveys/files/{filename}": {
      "get": {
        "operationId": "SurveysController_serveTempFile",
        "parameters": [{ "name": "filename", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["surveys"]
      }
    },
    "/edu-api/surveys/answers/files/{userName}/{surveyId}/{questionId}": {
      "post": {
        "operationId": "SurveysController_answeringFileUpload",
        "parameters": [
          { "name": "userName", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "surveyId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "questionId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["surveys"]
      }
    },
...
    "/edu-api/public-surveys/files/{surveyId}/{questionId}/{filename}": {
      "get": {
        "operationId": "PublicSurveysController_serveFile",
        "parameters": [
          { "name": "surveyId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "questionId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "filename", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["public-surveys"]
      }
    },
    "/edu-api/public-surveys/answers/files/{userName}/{surveyId}/{questionId}": {
      "post": {
        "operationId": "PublicSurveysController_answeringFileUpload",
        "parameters": [
          { "name": "userName", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "surveyId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "questionId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "201": { "description": "" } },
        "tags": ["public-surveys"]
      }
    },
    "/edu-api/public-surveys/answers/files/{userName}/{surveyId}/{questionId}/{fileName}": {
      "delete": {
        "operationId": "PublicSurveysController_deleteTempAnswerFile",
        "parameters": [
          { "name": "userName", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "surveyId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "questionId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "fileName", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["public-surveys"]
      },
      "get": {
        "operationId": "PublicSurveysController_serveFileFromAnswer",
        "parameters": [
          { "name": "userName", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "surveyId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "questionId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "fileName", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["public-surveys"]
      }
    },
...
    "/edu-api/bulletinboard/files": {
      "post": {
        "operationId": "BulletinBoardController_uploadBulletinAttachment",
        "parameters": [],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["bulletinboard"]
      }
    },
...
    "/edu-api/wireguard/peers/{peer}": {
      "delete": {
        "operationId": "WireguardController_deletePeer",
        "parameters": [{ "name": "peer", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Wireguard"]
      }
    },
    "/edu-api/wireguard/peers/{peer}/config": {
      "get": {
        "operationId": "WireguardController_getPeerConfig",
        "parameters": [{ "name": "peer", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Wireguard"]
      }
    },
    "/edu-api/wireguard/peers/{peer}/qr": {
      "get": {
        "operationId": "WireguardController_getPeerQR",
        "parameters": [{ "name": "peer", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Wireguard"]
      }
    },
    "/edu-api/wireguard/peers/{peer}/qr/b64": {
      "get": {
        "operationId": "WireguardController_getPeerQRBase64",
        "parameters": [{ "name": "peer", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Wireguard"]
      }
    },
    "/edu-api/wireguard/peers/{peer}/status": {
      "get": {
        "operationId": "WireguardController_getPeerStatus",
        "parameters": [{ "name": "peer", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Wireguard"]
      }
    },
...
    "/edu-api/wireguard/sites/{site}": {
      "delete": {
        "operationId": "WireguardController_deleteSite",
        "parameters": [{ "name": "site", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Wireguard"]
      }
    },
    "/edu-api/wireguard/sites/{site}/config": {
      "get": {
        "operationId": "WireguardController_getSiteConfig",
        "parameters": [{ "name": "site", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Wireguard"]
      }
    },
...
  },
  "info": { "title": "edulution-api", "description": "Test API for edulution-io", "version": "1.0", "contact": {} },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
...
  }
}